requirements.txt
Django>=4
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=3.13.0
drf-spectacular>=0.15.1
psycopg2>=2.8.6
redis>=4.4.0